Rename the module, expose the script as a command line, extract common system helper to a separate module.
