+++ src/scicat_ingestor.py
@@ -11,37 +11,10 @@ except importlib.metadata.PackageNotFoundError:
 
 del importlib
 
-import logging
-from collections.abc import Generator
-from contextlib import contextmanager
-
 from scicat_configuration import build_main_arg_parser, build_scicat_config
 from scicat_kafka import build_consumer, wrdn_messages
 from scicat_logging import build_logger
-
-
-def quit(logger: logging.Logger, unexpected: bool = True) -> None:
-    """Log the message and exit the program."""
-    import sys
-
-    logger.info("Exiting ingestor")
-    sys.exit(1 if unexpected else 0)
-
-
-@contextmanager
-def exit_at_exceptions(logger: logging.Logger) -> Generator[None, None, None]:
-    """Exit the program if an exception is raised."""
-    try:
-        yield
-    except KeyboardInterrupt:
-        logger.info("Received keyboard interrupt.")
-        quit(logger, unexpected=False)
-    except Exception as e:
-        logger.error("An exception occurred: %s", e)
-        quit(logger, unexpected=True)
-    else:
-        logger.error("Loop finished unexpectedly.")
-        quit(logger, unexpected=True)
+from system_helpers import exit_at_exceptions
 
 
 def main() -> None:
